Polishing

+++ file_management.bash
@@ -5,6 +5,10 @@ echo ------- Home Directories -------
 ls /home | column -c 85
 echo
 read -p "Username: " user
+if [[ $user == "" ]]; then
+	echo "No user selected"
+	exit
+fi
 path="/home/$user"
 if [[ "$(cat /etc/passwd | grep $user: | wc -l)" == 0 ]]; then
 	echo "User does not exist"
@@ -32,7 +36,12 @@ while true; do
 	case $input in
 		c | C)
 			read -p "input: " dir
-			if [[ "$(ls $path | grep $dir | wc -l)" == 0 ]]; then
+			if [[ $dir == "" ]]; then
+				echo "Invalid input"
+				bash exit.bash
+				exit
+			fi
+			if [[ "$(ls $path | grep -w $dir | wc -l)" == 0 ]]; then
 				echo "folder does not exist"
 				exit
 			fi
@@ -44,7 +53,12 @@ while true; do
 		v | V)
 			echo
 			read -p "input: " folderToView
-			if [[ "$(ls $path | grep $folderToView | wc -l)" == 0 ]]; then
+			if [[ $folderToView == "" ]]; then
+				echo "Invalid input"
+				bash exit.bash
+				exit
+			fi
+			if [[ "$(ls $path | grep -w $folderToView | wc -l)" == 0 ]]; then
 				echo "Folder does not exist"
 				exit
 			fi
@@ -87,6 +101,11 @@ while true; do
 			;;
 		a | A)
 			read -p "Name of new directory: " name
+			if [[ $name == "" ]]; then
+				echo "Name cant be empty"
+				bash exit.bash
+				exit
+			fi
 			mkdir $path/$name
 			gid=$(cat /etc/passwd | grep $user | cut ":"-f 4)
 			chown $user:$(cat /etc/group | grep -w $gid | cut -d ':' -f 1)  $path/$name
@@ -94,6 +113,14 @@ while true; do
 		d | D)
 			echo $path
 			read -p "Name of the directory you wish to remove: " name
+			if [[ $name == "" ]]; then
+				echo "Invalid input"
+				exit
+			fi 
+			if [[ $(ls $path | grep -w $name | wc -l) == 0 ]]; then
+				echo "directory does not exist"
+				exit
+			fi
 			read -p "Are you sure? (y/n): " choice
 			case $choice in
 				y | Y) rm -rf $path/$name ;;
